requirements.txt
bcrypt==4.0.1
blinker==1.6.2
click==8.1.7
distlib==0.3.7
dnspython==2.4.2
email-validator==2.0.0.post2
filelock==3.12.4
Flask==2.2.5
Flask-Bcrypt==1.0.1
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
idna==3.4
itsdangerous==2.0.1
Jinja2==3.1.2
MarkupSafe==2.1.3
Pillow==10.0.1
platformdirs==3.11.0
six==1.16.0
SQLAlchemy==2.0.21
typing_extensions==4.8.0
virtualenv==20.24.5
Werkzeug==2.3.0
WTForms==3.0.1
